Silence -Werror,-Wimplicit-int-float-conversion
> vcl/opengl/gdiimpl.cxx:2237:45: error: implicit conversion from 'int' to 'float' changes value from 2147483647 to 2147483648 [-Werror,-Wimplicit-int-float-conversion]
>     glClearColor(static_cast<float>(rand())/RAND_MAX, static_cast<float>(rand())/RAND_MAX,
>                                            ~^~~~~~~~

etc. with Clang 10, by doing the division with double precision.

+++ vcl/opengl/gdiimpl.cxx
@@ -2234,8 +2234,9 @@ void OpenGLSalGraphicsImpl::doFlush()
     mpWindowContext->state().stencil().disable();
 
 #if OSL_DEBUG_LEVEL > 0 // random background glClear
-    glClearColor(static_cast<float>(rand())/RAND_MAX, static_cast<float>(rand())/RAND_MAX,
-                 static_cast<float>(rand())/RAND_MAX, 1.0);
+    glClearColor(static_cast<float>(double(rand())/RAND_MAX),
+                 static_cast<float>(double(rand())/RAND_MAX),
+                 static_cast<float>(double(rand())/RAND_MAX), 1.0);
     glClear( GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT | GL_STENCIL_BUFFER_BIT );
     CHECK_GL_ERROR();
 #endif
